+++ routes/web.php
@@ -11,8 +11,11 @@ use Illuminate\Support\Facades\Route;
 | routes are loaded by the RouteServiceProvider within a group which
 | contains the "web" middleware group. Now create something great!
 |
-*/
+ */
 Route::get('/', function () {
+    return view('apps.website');
+});
+Route::get('/sign-file', function () {
     return view('apps.app');
 });
 Route::get('/signer', 'Signer\SignerController@index');